+++ .github/workflows/ci.yaml
@@ -79,4 +79,4 @@ jobs:
         uses: Swatinem/rust-cache@v1
 
       - name: Run tests
-        run: cargo test --verbose
+        run: cargo test --verbose --features parsnap
