+++ README.md
@@ -14,11 +14,35 @@ This crate provides a near drop in replacement for `Write` that has will compres
 to an underlying writer in the same order that the bytes were handed to the writer. This allows for much faster
 compression of Gzip data.
 
-## Usage
+### Supported Encodings:
+
+- Gzip via [flate2](https://docs.rs/flate2/)
+- Snappy via [rust-snappy](https://docs.rs/snap)
+
+## Usage / Features
+
+The default enabled features are "pargz" and "flate2_default" which enable gzip compression using whater flate2 uses as
+its default backend. To override this can do something like the folowing (choosing from available flate2 backends):
 
 ```toml
 [dependencies]
-gzp = { version = "*", features = ["zlib-ng-compat"] }
+gzp = { version = "*", no_default_features = true, features = ["pargz", "zlib-ng-compat"] }
+```
+
+To use "pargz" a backedn must be selected.
+
+To use Snap:
+
+```toml
+[dependencies]
+gzp = { version = "*", no_default_features = true, features = ["parsnap"] }
+```
+
+To use both Snap and Gzip
+
+```toml
+[dependencies]
+gzp = { version = "*", no_default_features = true, features = ["parsnap", "pargz", "zlib-ng-compat"] }
 ```
 
 ## Examples
@@ -28,7 +52,7 @@ Simple example
 ```rust
 use std::{env, fs::File, io::Write};
 
-use gzp::ParGz;
+use gzp::pargz::ParGz;
 
 fn main() {
     let file = env::args().skip(1).next().unwrap();
@@ -43,7 +67,7 @@ fn main() {
 An updated version of [pgz](https://github.com/vorner/pgz).
 
 ```rust
-use gzp::ParGz;
+use gzp::pargz::ParGz;
 use std::io::{Read, Write};
 
 fn main() {
@@ -69,10 +93,40 @@ fn main() {
 }
 ```
 
+Same thing but using Snappy instead.
+
+```rust
+use gzp::parsnap::ParSnap;
+use std::io::{Read, Write};
+
+fn main() {
+    let chunksize = 64 * (1 << 10) * 2;
+
+    let stdout = std::io::stdout();
+    let mut writer = ParSnap::builder(stdout).build();
+
+    let stdin = std::io::stdin();
+    let mut stdin = stdin.lock();
+
+    let mut buffer = Vec::with_capacity(chunksize);
+    loop {
+        let mut limit = (&mut stdin).take(chunksize as u64);
+        limit.read_to_end(&mut buffer).unwrap();
+        if buffer.is_empty() {
+            break;
+        }
+        writer.write_all(&buffer).unwrap();
+        buffer.clear();
+    }
+    writer.finish().unwrap();
+}
+```
+
 ## Notes
 
 - Files written with this are just Gzipped blocks catted together and must be read
   with `flate2::bufread::MultiGzDecoder`.
+
 
 ## Future todos
 
@@ -81,6 +135,25 @@ fn main() {
 - Update the CRC value with each block written
 - Add a BGZF mode + tabix index generation (or create that as its own crate)
 
-## Benchmarks
+## Benchmarks 
+
+All benchmarks were run on the file in `./bench-data/shakespeare.txt` catted together 100 times
+which creats a rough 550Mb file.
+
+| Name      | Num Threads | Compression Level | Buffer Size | Time | File Size | 
+| ---       | -           | ----------------- | ----------- | ---- | --------- |
+| Gzip Only | NA          | 3                 | 128 Kb      | 6.6s | 218 Mb    |
+| Gzip      | 1           | 3                 | 128 Kb      | 2.4s | 223 Mb    |
+| Gzip      | 4           | 3                 | 128 Kb      | 1.2s | 223 Mb    |
+| Gzip      | 8           | 3                 | 128 Kb      | 0.8s | 223 Mb    |
+| Gzip      | 16          | 3                 | 128 Kb      | 0.6s | 223 Mb    |
+| Gzip      | 30          | 3                 | 128 Kb      | 0.6s | 223 Mb    |
+| Snap Only | NA          | NA                | 128 Kb      | 1.6s | 333 Mb    |
+| Snap      | 1           | NA                | 128 Kb      | 0.7s | 333 Mb    |
+| Snap      | 4           | NA                | 128 Kb      | 0.5s | 333 Mb    |
+| Snap      | 8           | NA                | 128 Kb      | 0.4s | 333 Mb    |
+| Snap      | 16          | NA                | 128 Kb      | 0.4s | 333 Mb    |
+| Snap      | 30          | NA                | 128 Kb      | 0.4s | 333 Mb    |
+
 
 ![benchmarks](./violin.svg)
