+++ examples/snap_decode.rs
@@ -1,10 +1,22 @@
-use std::io;
+#[cfg(feature = "parsnap")]
+mod example {
+    use std::io;
+
+    pub fn main() {
+        let stdin = io::stdin();
+        let stdout = io::stdout();
+
+        let mut rdr = snap::read::FrameDecoder::new(stdin.lock());
+        let mut wtr = stdout.lock();
+        io::copy(&mut rdr, &mut wtr).expect("I/O operation failed");
+    }
+}
+
+#[cfg(not(feature = "parsnap"))]
+mod example {
+    pub fn main() {}
+}
 
 fn main() {
-    let stdin = io::stdin();
-    let stdout = io::stdout();
-
-    let mut rdr = snap::read::FrameDecoder::new(stdin.lock());
-    let mut wtr = stdout.lock();
-    io::copy(&mut rdr, &mut wtr).expect("I/O operation failed");
+    example::main()
 }
